pyproject.toml
[project]
name = "jsling"
version = "0.1.0"
description = "A remote job submission system for Slurm clusters"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fabric>=3.0",
    "click>=8.0",
    "sqlalchemy>=2.0",
    "cryptography>=41.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
jsling = "jsling.cli:main"
jsub = "jsling.commands.jsub:main"
jqueue = "jsling.commands.jqueue:main"
jsync = "jsling.commands.jsync:main"
jinfo = "jsling.commands.jinfo:main"
jcancel = "jsling.commands.jcancel:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
